Moved all references of player actions to the Board Information class
Removed class variables from the player class to the board information class as the separation is key to renaming the classes and their specific roles. In future the Player class should be a model container, and the board information class should handle all the data that relates to the players on the board instead. In terms of abstraction the model container is now "full agent" for the whole game where each model that it possesses are the agents for the specific action. The board information class will in future just be called board. as it should in theory contain all the information that the board can possess and all the methods that control the state of the board